refactor: merge organize service files into internal/organizer
Move organize_service.go, organize_preview_service.go, rename_service.go,
file_pipeline.go, file_move.go, and path_format.go into the internal/organizer
package. Server-specific dependencies (isProtectedPath, iTunes sync,
WriteBackBatcher) are injected via callback fields to avoid import cycles.
Original server files become thin forwarding layers with type aliases.
